Support BAMCOV BigWig path export without implicit hacks.
The extension lookup has been modified to hold a 2 level hash. The first
is keyed by the DataFile type and the second is keyed by the requested
file type. In most cases these are the same. In BAMCOV we have 3
sub-hashes: BAM, BAMCOV and BIGWIG all of which return the right
extension required.